+++ CHANGES.md
@@ -32,6 +32,8 @@ Unreleased
   - The specific methods haven't changed (to make it easy to migrate, they might be revisited later) but they have become actual method instead of staticmethods.
   - The `Executor.supports_subscriptions` attribute has been replaced by the `SubscriptionRuntime` subclass.
 
+- `SchemaVisitor.(on_field_definition|on_argument_definition|on_input_field_definition)` have become `SchemaVisitor.(on_field|on_argument|on_input_field)` to maintain consistency with other methods.
+
 [0.4.0](https://github.com/lirsacc/py-gql/releases/tag/0.4.0) - 2019-10-10
 --------------------------------------------------------------------------